Split UI tests into seprate files.
